Renaming balanceUseCase
