+++ app/src/main/java/br/com/littlepig/di/UseCaseModule.kt
@@ -1,11 +1,15 @@
 package br.com.littlepig.di
 
+import br.com.littlepig.domain.usecase.balance.BalanceUseCase
+import br.com.littlepig.domain.usecase.balance.IBalanceUseCase
 import br.com.littlepig.domain.usecase.login.ILoginUseCase
 import br.com.littlepig.domain.usecase.login.LoginUseCase
 import br.com.littlepig.domain.usecase.register.IRegisterUseCase
 import br.com.littlepig.domain.usecase.register.RegisterUseCase
-import br.com.littlepig.domain.usecase.transactions.ITransactionsUseCase
-import br.com.littlepig.domain.usecase.transactions.TransactionsUseCase
+import br.com.littlepig.domain.usecase.transactions.DeleteTransactionUseCase
+import br.com.littlepig.domain.usecase.transactions.GetTransactionsUseCase
+import br.com.littlepig.domain.usecase.transactions.IDeleteTransactionUseCase
+import br.com.littlepig.domain.usecase.transactions.IGetTransactionsUseCase
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -21,5 +25,11 @@ interface UseCaseModule {
     fun provideLoginUseCase(loginUseCase: LoginUseCase): ILoginUseCase
 
     @Binds
-    fun provideTransactionsUseCase(transactionsUseCase: TransactionsUseCase): ITransactionsUseCase
+    fun provideBalanceUseCase(balanceUseCase: BalanceUseCase): IBalanceUseCase
+
+    @Binds
+    fun provideTransactionsUseCase(getTransactionsUseCase: GetTransactionsUseCase): IGetTransactionsUseCase
+
+    @Binds
+    fun provideDeleteTransactionUseCase(deleteTransactionUseCase: DeleteTransactionUseCase): IDeleteTransactionUseCase
 }
