+++ app/src/main/java/br/com/littlepig/domain/usecase/balance/IBalanceUseCase.kt
@@ -1,7 +1,7 @@
-package br.com.littlepig.domain.usecase.transactions
+package br.com.littlepig.domain.usecase.balance
 
 import br.com.littlepig.data.model.balance.UserBalanceResponseItem
 
-interface ITransactionsUseCase {
+interface IBalanceUseCase {
     suspend operator fun invoke(date: Long): Result<List<UserBalanceResponseItem>>
 }
